time-series-mixup-src/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.64.0
pyyaml>=6.0
psutil>=5.9.0
sktime>=0.16.0
aeon>=0.3.0
gdown>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0